Separate code into mobile and wear packages
