+++ mobile/src/main/AndroidManifest.xml
@@ -12,18 +12,18 @@
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
-        android:name=".model.GlobalClient"
+        android:name=".mobile.model.GlobalClient"
         android:theme="@style/AppTheme">
-        <activity android:name=".Home">
+        <activity android:name=".mobile.activity.Home">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".Monitoring" />
-        <activity android:name=".PreDive" />
-        <activity android:name=".DiveInProgress"></activity>
+        <activity android:name=".mobile.activity.Monitoring" />
+        <activity android:name=".mobile.activity.PreDive" />
+        <activity android:name=".mobile.activity.DiveInProgress"></activity>
     </application>
 
 </manifest>
