+++ mobile/src/main/java/com/example/zmeggyesi/divemonitor/mobile/model/EnvironmentReading.java
@@ -1,4 +1,4 @@
-package com.example.zmeggyesi.divemonitor.model;
+package com.example.zmeggyesi.divemonitor.mobile.model;
 
 import android.provider.BaseColumns;
