+++ mobile/src/main/java/com/example/zmeggyesi/divemonitor/mobile/service/DiveDatabaseHelper.java
@@ -1,11 +1,10 @@
-package com.example.zmeggyesi.divemonitor.services;
+package com.example.zmeggyesi.divemonitor.mobile.service;
 
 import android.content.Context;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.util.Log;
 
-import com.example.zmeggyesi.divemonitor.model.Dive.Record;
+import com.example.zmeggyesi.divemonitor.mobile.model.Dive.Record;
 
 /**
  * Created by zmeggyesi on 2017. 03. 15..
