+++ wear/src/main/AndroidManifest.xml
@@ -11,7 +11,7 @@
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
-        android:name=".model.GlobalContext"
+        android:name=".wear.model.GlobalContext"
         android:launchMode="singleInstance"
         android:theme="@android:style/Theme.DeviceDefault">
         <uses-library
@@ -19,7 +19,7 @@
             android:required="false"/>
 
         <activity
-            android:name=".Monitor"
+            android:name=".wear.activity.Monitor"
             android:label="@string/app_name"
             android:theme="@android:style/Theme.DeviceDefault.Light">
             <intent-filter>
@@ -29,7 +29,7 @@
             </intent-filter>
         </activity>
 
-        <service android:name=".MonitorTriggerListener">
+        <service android:name=".wear.activity.MonitorTriggerListener">
             <intent-filter>
                 <action android:name="com.google.android.gms.wearable.MESSAGE_RECEIVED"/>
 
@@ -46,7 +46,7 @@
             </intent-filter>
         </service>
         <service
-            android:name=".services.RecorderService"
+            android:name=".wear.services.RecorderService"
             android:enabled="true"
             android:exported="false">
         </service>
