+++ wear/src/main/java/com/example/zmeggyesi/divemonitor/wear/model/GlobalContext.java
@@ -1,39 +1,31 @@
-package com.example.zmeggyesi.divemonitor.model;
+package com.example.zmeggyesi.divemonitor.wear.model;
 
 import android.app.Application;
+import android.content.Context;
 import android.content.Intent;
 import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
+import android.hardware.SensorManager;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
-import com.example.zmeggyesi.divemonitor.services.DiveDatabaseHelper;
-import com.example.zmeggyesi.divemonitor.services.EnvironmentReadingDatabaseHelper;
+import com.example.zmeggyesi.divemonitor.wear.services.RemoteEnvironmentDatabaseHelper;
 import com.google.android.gms.common.ConnectionResult;
-import com.google.android.gms.common.api.GoogleApiActivity;
 import com.google.android.gms.common.api.GoogleApiClient;
-import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.tasks.RuntimeExecutionException;
 import com.google.android.gms.wearable.Wearable;
 
 /**
  * Created by zmeggyesi on 2017. 03. 07..
  */
 
-public class GlobalClient extends Application implements GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
+public class GlobalContext extends Application implements GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
 	private final String TAG = "Global Context";
 	private GoogleApiClient apiClient;
-	private DiveDatabaseHelper divesHelper;
-	private EnvironmentReadingDatabaseHelper environmentReadingsHelper;
+	private RemoteEnvironmentDatabaseHelper remoteEnvironmentDatabaseHelper;
 
-	@Override
-	public void onTerminate() {
-		divesHelper.close();
-		environmentReadingsHelper.close();
-		apiClient.disconnect();
-		super.onTerminate();
+	public SensorManager getSensorManager() {
+		return (SensorManager) getSystemService(Context.SENSOR_SERVICE);
 	}
 
 	@Override
@@ -41,7 +33,6 @@ public class GlobalClient extends Application implements GoogleApiClient.Connect
 		super.onCreate();
 		apiClient = new GoogleApiClient.Builder(this)
 				.addApi(Wearable.API)
-				.addApi(LocationServices.API)
 				.addConnectionCallbacks(this)
 				.addOnConnectionFailedListener(this)
 				.build();
@@ -53,17 +44,11 @@ public class GlobalClient extends Application implements GoogleApiClient.Connect
 		if (apiClient == null) {
 			return apiClient = new GoogleApiClient.Builder(this)
 					.addApi(Wearable.API)
-					.addApi(LocationServices.API)
 					.addConnectionCallbacks(this)
 					.addOnConnectionFailedListener(this)
 					.build();
 		} else {
-			if (apiClient.isConnected()) {
-				return apiClient;
-			} else {
-				apiClient.connect();
-				return apiClient;
-			}
+			return apiClient;
 		}
 	}
 
@@ -86,28 +71,16 @@ public class GlobalClient extends Application implements GoogleApiClient.Connect
 	}
 
 	private void setupDBs() {
-		divesHelper = new DiveDatabaseHelper(this);
-		environmentReadingsHelper = new EnvironmentReadingDatabaseHelper(this);
+		remoteEnvironmentDatabaseHelper = new RemoteEnvironmentDatabaseHelper(this);
 	}
 
 	public SQLiteDatabase getEnvironmentReadingsDatabase(boolean rw) {
 		if (rw) {
-			SQLiteDatabase writableDatabase = environmentReadingsHelper.getWritableDatabase();
+			SQLiteDatabase writableDatabase = remoteEnvironmentDatabaseHelper.getWritableDatabase();
 			writableDatabase.setForeignKeyConstraintsEnabled(true);
 			return writableDatabase;
 		} else {
-			SQLiteDatabase readableDatabase = environmentReadingsHelper.getReadableDatabase();
-			return readableDatabase;
-		}
-	}
-
-	public SQLiteDatabase getDivesDatabase(boolean rw) {
-		if (rw) {
-			SQLiteDatabase writableDatabase = divesHelper.getWritableDatabase();
-			writableDatabase.setForeignKeyConstraintsEnabled(true);
-			return writableDatabase;
-		} else {
-			SQLiteDatabase readableDatabase = divesHelper.getReadableDatabase();
+			SQLiteDatabase readableDatabase = remoteEnvironmentDatabaseHelper.getReadableDatabase();
 			return readableDatabase;
 		}
 	}
