+++ wear/src/main/java/com/example/zmeggyesi/divemonitor/wear/sensorium/OrientationHandler.java
@@ -1,4 +1,4 @@
-package com.example.zmeggyesi.divemonitor.sensorium;
+package com.example.zmeggyesi.divemonitor.wear.sensorium;
 
 import android.content.ComponentName;
 import android.content.Context;
@@ -12,9 +12,9 @@ import android.os.IBinder;
 import android.support.v4.content.LocalBroadcastManager;
 import android.util.Log;
 
-import com.example.zmeggyesi.divemonitor.Monitor;
+import com.example.zmeggyesi.divemonitor.wear.activity.Monitor;
 import com.example.zmeggyesi.divemonitor.R;
-import com.example.zmeggyesi.divemonitor.services.RecorderService;
+import com.example.zmeggyesi.divemonitor.wear.services.RecorderService;
 
 import java.util.ArrayList;
 import java.util.List;
