+++ wear/src/main/res/layout/activity_monitor.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.example.zmeggyesi.divemonitor.Monitor"
+    tools:context="com.example.zmeggyesi.divemonitor.wear.activity.Monitor"
     tools:deviceIds="wear">
 
     <LinearLayout
